feat: 팀 삭제에 대한 로직 수정, 서비스 추가
- bulk 쿼리에 대한 트랜잭션 분리
- 리더 변경, 멤버 추방 메소드 추가

+++ team-service/src/main/java/com/teamservice/application/TeamService.java
@@ -2,6 +2,7 @@ package com.teamservice.application;
 
 import com.teamservice.application.exception.AlreadyCreatedException;
 import com.teamservice.application.exception.DeleteNotAllowedException;
+import com.teamservice.application.exception.NotBelongToTeamException;
 import com.teamservice.application.exception.NotLeaderException;
 import com.teamservice.application.exception.TeamNameDuplicateException;
 import com.teamservice.application.exception.TeamNotFoundException;
@@ -10,8 +11,10 @@ import com.teamservice.dto.query.TeamQueryDto;
 import com.teamservice.dto.request.TeamRequestDto;
 import com.teamservice.dto.request.TeamUpdateRequestDto;
 import com.teamservice.persistence.Team;
+import com.teamservice.persistence.TeamUser;
 import com.teamservice.persistence.repository.RequestHistoryRepository;
 import com.teamservice.persistence.repository.TeamRepository;
+import com.teamservice.persistence.repository.TeamUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,6 +24,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class TeamService {
 
     private final TeamRepository teamRepository;
+    private final TeamUserRepository teamUserRepository;
     private final RequestHistoryRepository requestHistoryRepository;
 
     @Transactional
@@ -38,16 +42,33 @@ public class TeamService {
     }
 
     @Transactional
-    public void deleteTeam(Long teamId, String userId) {
+    public void changeLeader(Long teamId, String userId, String leaderId) {
         Team team = findTeam(teamId);
-        validateOnDeleteTeam(userId, team);
-        teamRepository.delete(team); // cascade delete: TeamUser list
-        deleteRequestListByTeamId(team.getId()); // bulk
+        validateOnUpdateTeam(userId, team);
+        team.changeLeader(leaderId);
     }
 
     @Transactional
-    protected void deleteRequestListByTeamId(Long id) {
+    public void deleteTeam(Long teamId, String userId) {
+        Team team = teamRepository.findWithTeamUserById(teamId);
+        validateOnDeleteTeam(userId, team);
+        teamUserRepository.deleteAllInBatch(team.getTeamUserList());
+        teamRepository.deleteById(teamId);
+    }
+
+    @Transactional
+    public void deleteRequestListByTeamId(Long id) {
+        if (!teamRepository.existsById(id)) throw new TeamNotFoundException();
         requestHistoryRepository.deleteListByTeamId(id);
+    }
+
+    @Transactional
+    public void withdrawMember(Long teamId, String leaderId, String userId) {
+        Team team = teamRepository.getReferenceById(teamId);
+        validateOnUpdateTeam(leaderId, team);
+        TeamUser teamUser = teamUserRepository.findByTeamAndUserId(team, userId)
+                .orElseThrow(NotBelongToTeamException::new);
+        teamUserRepository.delete(teamUser);
     }
 
     @Transactional(readOnly = true)
